feat: solution module and hash tasks
