+++ src/test.js
@@ -1,0 +1,4 @@
+import { Buffer } from 'node:buffer'
+
+const emptyBuffer = Buffer.alloc(42, 1)
+console.log(emptyBuffer)
